+++ libs/numeric/ublasx/doc/Changes.md
@@ -9,7 +9,7 @@ For a more detailed list, please see the commit log (e.g., go to the GitHub [his
 ### Breaking Changes
 
 - Dropped support for C++ standards less than C++11.
-- Changed the following operations to work like their MATLAB/Octave counterparts: `isinf`, `sign`.
+- Changed the following operations to work like their MATLAB/Octave counterparts: `isinf`, `reshape`, `sign`.
 
 ### New Features